+++ top_speed_net/TopSpeed.Shared/TrackTypes.cs
@@ -58,9 +58,9 @@ namespace TopSpeed.Data
         public TrackType Type { get; }
         public TrackSurface Surface { get; }
         public TrackNoise Noise { get; }
-        public int Length { get; }
+        public float Length { get; }
 
-        public TrackDefinition(TrackType type, TrackSurface surface, TrackNoise noise, int length)
+        public TrackDefinition(TrackType type, TrackSurface surface, TrackNoise noise, float length)
         {
             Type = type;
             Surface = surface;
