+++ top_speed_net/TopSpeed/Vehicles/VehicleDefinition.cs
@@ -7,14 +7,14 @@ namespace TopSpeed.Vehicles
         public CarType CarType { get; set; }
         public bool UserDefined { get; set; }
         public string? CustomFile { get; set; }
-        public int Acceleration { get; set; }
-        public int Deceleration { get; set; }
-        public int TopSpeed { get; set; }
+        public float Acceleration { get; set; }
+        public float Deceleration { get; set; }
+        public float TopSpeed { get; set; }
         public int IdleFreq { get; set; }
         public int TopFreq { get; set; }
         public int ShiftFreq { get; set; }
         public int Gears { get; set; }
-        public int Steering { get; set; }
+        public float Steering { get; set; }
         public int SteeringFactor { get; set; }
         public int HasWipers { get; set; }
         public string? EngineSound { get; set; }
